+++ astropy/timeseries/periodograms/__init__.py
@@ -1,0 +1,2 @@
+from astropy.timeseries.periodograms.lombscargle import *  # noqa
+from astropy.timeseries.periodograms.bls import *  # noqa
