+++ astropy/timeseries/periodograms/lombscargle/__init__.py
