+++ astropy/timeseries/periodograms/lombscargle/_statistics.py
