+++ astropy/timeseries/periodograms/lombscargle/core.py
