+++ astropy/timeseries/periodograms/lombscargle/implementations/chi2_impl.py
