+++ astropy/timeseries/periodograms/lombscargle/implementations/fast_impl.py
