+++ astropy/timeseries/periodograms/lombscargle/implementations/fastchi2_impl.py
