+++ astropy/timeseries/periodograms/lombscargle/implementations/scipy_impl.py
