+++ astropy/timeseries/periodograms/lombscargle/implementations/slow_impl.py
