+++ astropy/timeseries/periodograms/lombscargle/implementations/tests/test_mle.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose
 
-from astropy.timeseries.lombscargle.implementations.mle import design_matrix, periodic_fit
+from astropy.timeseries.periodograms.lombscargle.implementations.mle import design_matrix, periodic_fit
 
 
 @pytest.fixture
