+++ astropy/timeseries/periodograms/lombscargle/implementations/tests/test_utils.py
@@ -3,7 +3,7 @@ import pytest
 import numpy as np
 from numpy.testing import assert_allclose, assert_equal
 
-from astropy.timeseries.lombscargle.implementations.utils import extirpolate, bitceil, trig_sum
+from astropy.timeseries.periodograms.lombscargle.implementations.utils import extirpolate, bitceil, trig_sum
 
 
 @pytest.mark.parametrize('N', 2 ** np.arange(1, 12))
