+++ astropy/timeseries/periodograms/lombscargle/implementations/utils.py
