+++ astropy/timeseries/periodograms/lombscargle/tests/test_statistics.py
@@ -9,10 +9,10 @@ except ImportError:
 else:
     HAS_SCIPY = True
 
-from astropy.timeseries.lombscargle import LombScargle
-from astropy.timeseries.lombscargle._statistics import (cdf_single, pdf_single, fap_single, inv_fap_single,
+from astropy.timeseries.periodograms.lombscargle import LombScargle
+from astropy.timeseries.periodograms.lombscargle._statistics import (cdf_single, pdf_single, fap_single, inv_fap_single,
                            METHODS)
-from astropy.timeseries.lombscargle.utils import convert_normalization, compute_chi2_ref
+from astropy.timeseries.periodograms.lombscargle.utils import convert_normalization, compute_chi2_ref
 
 METHOD_KWDS = dict(bootstrap={'n_bootstraps': 20, 'random_seed': 42})
 NORMALIZATIONS = ['standard', 'psd', 'log', 'model']
