+++ astropy/timeseries/periodograms/lombscargle/tests/test_utils.py
@@ -2,8 +2,8 @@ import numpy as np
 import pytest
 from numpy.testing import assert_allclose
 
-from astropy.timeseries.lombscargle.utils import convert_normalization, compute_chi2_ref
-from astropy.timeseries.lombscargle.core import LombScargle
+from astropy.timeseries.periodograms.lombscargle.utils import convert_normalization, compute_chi2_ref
+from astropy.timeseries.periodograms.lombscargle.core import LombScargle
 
 
 NORMALIZATIONS = ['standard', 'model', 'log', 'psd']
